+++ integration/js/ContentShareOnlyAllowTwoTest.js
@@ -39,30 +39,30 @@ class ContentShareOnlyAllowTwoTest extends SdkBaseTest {
     await test_window_1.runCommands(async () => await ClickContentShareButton.executeStep(this, session, "ON"));
     await TestUtils.waitAround(5000);
     await test_window_1.runCommands(async () => await RosterCheck.executeStep(this, session, 4));
-    await test_window_2.runCommands(async () => await ContentShareVideoCheck.executeStep(this, session, "ON", 2));
+    await test_window_2.runCommands(async () => await ContentShareVideoCheck.executeStep(this, session, "ON", 1));
     await test_window_2.runCommands(async () => await RosterCheck.executeStep(this, session, 4));
-    await test_window_3.runCommands(async () => await ContentShareVideoCheck.executeStep(this, session, "ON", 2));
+    await test_window_3.runCommands(async () => await ContentShareVideoCheck.executeStep(this, session, "ON", 1));
     await test_window_3.runCommands(async () => await RosterCheck.executeStep(this, session, 4));
 
     //Turn on Content Share for second participant
     await test_window_2.runCommands(async () => await ClickContentShareButton.executeStep(this, session, "ON"));
     await TestUtils.waitAround(5000);
     await test_window_2.runCommands(async () => await RosterCheck.executeStep(this, session, 5));
-    await test_window_1.runCommands(async () => await ContentShareVideoCheck.executeStep(this, session, "ON", 2));
+    await test_window_1.runCommands(async () => await ContentShareVideoCheck.executeStep(this, session, "ON", 1));
     await test_window_1.runCommands(async () => await RosterCheck.executeStep(this, session, 5));
+    await test_window_3.runCommands(async () => await ContentShareVideoCheck.executeStep(this, session, "ON", 1));
     await test_window_3.runCommands(async () => await ContentShareVideoCheck.executeStep(this, session, "ON", 2));
-    await test_window_3.runCommands(async () => await ContentShareVideoCheck.executeStep(this, session, "ON", 3));
     await test_window_3.runCommands(async () => await RosterCheck.executeStep(this, session, 5));
 
     //Turn on Content Share for third participant
     await test_window_3.runCommands(async () => await ClickContentShareButton.executeStep(this, session, "ON"));
     await TestUtils.waitAround(5000);
+    await test_window_3.runCommands(async () => await ContentShareVideoCheck.executeStep(this, session, "ON", 1));
     await test_window_3.runCommands(async () => await ContentShareVideoCheck.executeStep(this, session, "ON", 2));
-    await test_window_3.runCommands(async () => await ContentShareVideoCheck.executeStep(this, session, "ON", 3));
     await test_window_3.runCommands(async () => await RosterCheck.executeStep(this, session, 5));
-    await test_window_1.runCommands(async () => await ContentShareVideoCheck.executeStep(this, session, "ON", 2));
+    await test_window_1.runCommands(async () => await ContentShareVideoCheck.executeStep(this, session, "ON", 1));
     await test_window_1.runCommands(async () => await RosterCheck.executeStep(this, session, 5));
-    await test_window_2.runCommands(async () => await ContentShareVideoCheck.executeStep(this, session, "ON", 2));
+    await test_window_2.runCommands(async () => await ContentShareVideoCheck.executeStep(this, session, "ON", 1));
     await test_window_2.runCommands(async () => await RosterCheck.executeStep(this, session, 5));
   }
 }