+++ integration/js/MeetingLeaveContentShareTest.js
@@ -29,12 +29,12 @@ class MeetingLeaveContentShareTest extends SdkBaseTest {
     await TestUtils.waitAround(5000);
     await test_window.runCommands(async () => await RosterCheck.executeStep(this, session, 3));
     await monitor_window.runCommands(async () => await RosterCheck.executeStep(this, session, 3));
-    await monitor_window.runCommands(async () => await ContentShareVideoCheck.executeStep(this, session, "ON", 2));
+    await monitor_window.runCommands(async () => await ContentShareVideoCheck.executeStep(this, session, "ON", 1));
 
     await test_window.runCommands(async () => await LeaveMeetingStep.executeStep(this, session));
     await TestUtils.waitAround(5000);
     await monitor_window.runCommands(async () => await RosterCheck.executeStep(this, session, 1));
-    await monitor_window.runCommands(async () => await ContentShareVideoCheck.executeStep(this, session, "OFF", 2));
+    await monitor_window.runCommands(async () => await ContentShareVideoCheck.executeStep(this, session, "OFF", 1));
   }
 }